splits FrejaeID into separate component and adds idp from config
